requirements.txt
yfinance==1.3.0
python-telegram-bot==22.7
python-dotenv==1.0.0
ta==0.11.0
schedule==1.2.2